+++ ndonnx/_core/_impl.py
@@ -199,12 +199,10 @@ class CoreOperationsImpl(OperationsBlock):
         return self.add(self.log(x), ndx.asarray(1, x.dtype))
 
     def log2(self, x):
-        out = ndx.log(x)
-        return out / ndx.asarray(np.log(2), dtype=out.dtype)
+        return ndx.log(x) / np.log(2)
 
     def log10(self, x):
-        out = ndx.log(x)
-        return out / ndx.asarray(np.log(10), dtype=out.dtype)
+        return ndx.log(x) / np.log(10)
 
     def logaddexp(self, x, y):
         return self.log(self.exp(x) + self.exp(y))